Create README.md
In this project we will assume that we have recently been hired as a Data Analyst by a global IT and business consulting services firm that is known for their expertise in IT solutions and their team of highly experienced IT consultants. In order to keep pace with changing technologies and remain competitive, our organization regularly analyzes data to help identify future skill requirements.

As a Data Analyst, we will be assisting with this initiative and have been tasked with collecting data from various sources and identifying trends for this year's report on emerging skills.

Our first task is to collect the top programming skills that are most in demand from various sources including: Job postings / Training portals / Surveys

Once we have collected enough data, we will begin analyzing the data and identify insights and trends.

Each step of the data analysis process is shown in different notebook of this repository and some additional files needed for their understanding.